+++ deneb/config.py
@@ -1,9 +1,9 @@
 """Module to store app config"""
 
-VERSION = 'v2.0.4'
+VERSION = "v2.1.1"
 
 
 class Config:
     USERS_TASKS_AMOUNT = 5
-    ALBUMS_TASKS_AMOUNT = 50
-    ARTISTS_TASKS_AMOUNT = 30
+    ALBUMS_TASKS_AMOUNT = 20
+    ARTISTS_TASKS_AMOUNT = 20